requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.0
matplotlib==3.8.0
